+++ Core/Audio/Impl/Components/OpenALBuffer.cs
@@ -1,5 +1,6 @@
-using System;
 using OpenTK.Audio.OpenAL;
+using System;
+using ZMusicWrapper;
 using static Helion.Audio.AudioHelper;
 using static Helion.Util.Assertion.Assert;
 
@@ -29,14 +30,70 @@ public class OpenALBuffer : IDisposable
     public static OpenALBuffer? Create(byte[] data, out string? error)
     {
         error = null;
-        if (IsWave(data) && TryReadWav(data, out WavFormat wavFormat, out Span<byte> sampleData) && GetFormat(wavFormat, out ALFormat format, out error))
-            return new OpenALBuffer(wavFormat.SampleRate, sampleData, format);
-        else if (IsDoomSound(data) && TryReadDoomSound(data, out int sampleRate, out sampleData, out error))
-            return new OpenALBuffer(sampleRate, sampleData, ALFormat.Mono8);
-        else if (error == null)
-            error = "Unsupported format.";
+        if (data.Length == 0)
+            return null;
 
+        if (IsWave(data) && TryReadWav(data, out var wavFormat, out var sampleData) && GetFormat(wavFormat, out var format, out error))
+        {
+            return new(wavFormat.SampleRate, sampleData, format);
+        }
+        else if (IsDoomSound(data) && TryReadDoomSound(data, out var sampleRate, out sampleData, out error))
+        {
+            return new(sampleRate, sampleData, ALFormat.Mono8);
+        }
+        else 
+        if (TryConvertThroughZMusic(data, out sampleRate, out var stereo, out var sampleType, out var convertedData) &&
+            GetFormat(stereo, sampleType, out format, out error))
+        {
+            return new(sampleRate, convertedData, format);
+        }
+
+        error ??= "Unsupported format.";
         return null;
+    }
+
+    private static bool GetFormat(bool stereo, SampleType sampleType, out ALFormat format, out string? error)
+    {
+        format = 0;
+        if (stereo)
+        {
+            switch (sampleType)
+            {
+                case SampleType.UInt8:
+                    format = ALFormat.Stereo8;
+                    break;
+                case SampleType.Int16:
+                    format = ALFormat.Stereo16;
+                    break;
+                case SampleType.Float32:
+                    format = ALFormat.StereoFloat32Ext;
+                    break;
+            }
+        }
+        else
+        {
+            switch (sampleType)
+            {
+                case SampleType.UInt8:
+                    format = ALFormat.Mono8;
+                    break;
+                case SampleType.Int16:
+                    format = ALFormat.Mono16;
+                    break;
+                case SampleType.Float32:
+                    format = ALFormat.MonoFloat32Ext;
+                    break;
+            }
+        }
+
+        if (format == 0)
+        {
+            error = $"Unsupported format stereo:{stereo} sample:{sampleType}";
+            return false;
+        }
+
+        error = null;
+        return true;
     }
 
     private static bool GetFormat(WavFormat wavFormat, out ALFormat format, out string? error)